requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alpha-vantage==2.3.1
altair==5.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
feedparser==6.0.11
frozendict==2.4.4
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
html5lib==1.1
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
multitasking==0.0.11
nltk==3.8.1
numpy==1.26.4
packaging==24.0
pandas==2.2.2
patsy==0.5.6
peewee==3.17.5
pillow==10.3.0
platformdirs==4.2.2
plotly==5.22.0
plotly-express==0.4.1
protobuf==4.25.3
pyarrow==16.1.0
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
scipy==1.13.1
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.1
soupsieve==2.5
statsmodels==0.14.2
stocknews==0.9.11
streamlit==1.35.0
tenacity==8.3.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.1
webencodings==0.5.1
yarl==1.9.4
yfinance==0.2.40
